Fix file names to prevent MacOS casing issues (#2534)
