+++ src/test/java/org/assertj/core/api/double_/DoubleAssert_isNotEqualTo_double_primitive_Test.java
@@ -10,11 +10,12 @@
  *
  * Copyright 2012-2022 the original author or authors.
  */
-package org.assertj.core.api.float_;
+package org.assertj.core.api.double_;
 
 import static java.lang.String.format;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.BDDAssertions.then;
+import static org.assertj.core.error.ShouldNotBeNull.shouldNotBeNull;
 import static org.assertj.core.test.AlwaysDifferentComparator.ALWAY_DIFFERENT;
 import static org.assertj.core.util.AssertionsUtil.expectAssertionError;
 import static org.mockito.BDDMockito.given;
@@ -23,60 +24,71 @@ import static org.mockito.Mockito.verifyNoMoreInteractions;
 
 import java.util.Comparator;
 
-import org.assertj.core.api.FloatAssert;
-import org.assertj.core.api.FloatAssertBaseTest;
+import org.assertj.core.api.DoubleAssert;
+import org.assertj.core.api.DoubleAssertBaseTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 /**
- * Tests for <code>{@link FloatAssert#isNotEqualTo(float)}</code>.
+ * Tests for <code>{@link DoubleAssert#isNotEqualTo(double)}</code>.
  *
  * @author Alex Ruiz
  */
-class FloatAssert_isNotEqualTo_float_Test extends FloatAssertBaseTest {
+class DoubleAssert_isNotEqualTo_double_primitive_Test extends DoubleAssertBaseTest {
 
   @SuppressWarnings({ "unchecked", "rawtypes" })
   @Override
-  protected FloatAssert invoke_api_method() {
+  protected DoubleAssert invoke_api_method() {
     // trick to simulate a custom comparator
-    given(floats.getComparator()).willReturn((Comparator) ALWAY_EQUAL_FLOAT);
-    return assertions.isNotEqualTo(8f);
+    given(doubles.getComparator()).willReturn((Comparator) ALWAY_EQUAL_DOUBLE);
+    return assertions.isNotEqualTo(8d);
   }
 
   @Override
   protected void verify_internal_effects() {
-    verify(floats).getComparator();
-    verify(floats).assertNotEqual(getInfo(assertions), getActual(assertions), 8f);
-    verifyNoMoreInteractions(floats);
+    verify(doubles).getComparator();
+    verify(doubles).assertNotEqual(getInfo(assertions), getActual(assertions), 8d);
+    verifyNoMoreInteractions(doubles);
   }
 
   @ParameterizedTest
-  @CsvSource({ "1.0f, -1.0f", "NaN, NaN" })
-  void should_pass_using_primitive_comparison(float actual, float expected) {
+  @CsvSource({ "1.0, -1.0", "NaN, NaN" })
+  void should_pass_using_primitive_comparison(double actual, double expected) {
     assertThat(actual).isNotEqualTo(expected);
   }
 
   @Test
   void should_honor_user_specified_comparator() {
     // GIVEN
-    final float one = 1.0f;
+    final double one = 1.0d;
     // THEN
     assertThat(one).usingComparator(ALWAY_DIFFERENT)
                    .isNotEqualTo(one);
   }
 
   @Test
-  void should_fail_if_floats_are_equal() {
+  void should_fail_if_doubles_are_equal() {
     // GIVEN
-    float actual = 0.0f;
-    float expected = -0.0f;
+    double actual = 0.0;
+    double expected = -0.0;
     // WHEN
     AssertionError assertionError = expectAssertionError(() -> assertThat(actual).isNotEqualTo(expected));
     // THEN
     then(assertionError).hasMessage(format("%nExpecting actual:%n" +
-                                           "  0.0f%n" +
+                                           "  0.0%n" +
                                            "not to be equal to:%n" +
-                                           "  -0.0f%n"));
+                                           "  -0.0%n"));
+  }
+
+  @Test
+  void should_fail_when_actual_null_expected_primitive() {
+    // GIVEN
+    Double actual = null;
+    double expected = 1.0d;
+    // WHEN
+    AssertionError assertionError = expectAssertionError(() -> assertThat(actual).isNotEqualTo(expected));
+    // THEN
+    then(assertionError).hasMessageContaining(shouldNotBeNull().create());
   }
 }
