+++ src/test/java/org/assertj/core/api/float_/FloatAssert_isNotCloseToPercentage_float_primitive_Test.java
@@ -10,32 +10,32 @@
  *
  * Copyright 2012-2022 the original author or authors.
  */
-package org.assertj.core.api.double_;
+package org.assertj.core.api.float_;
 
 import static org.assertj.core.data.Percentage.withPercentage;
 import static org.mockito.Mockito.verify;
 
-import org.assertj.core.api.DoubleAssert;
-import org.assertj.core.api.DoubleAssertBaseTest;
+import org.assertj.core.api.FloatAssert;
+import org.assertj.core.api.FloatAssertBaseTest;
 import org.assertj.core.data.Percentage;
 
 /**
- * Tests for <code>{@link DoubleAssert#isNotCloseTo(double, Percentage)}</code>.
+ * Tests for <code>{@link FloatAssert#isNotCloseTo(float, Percentage)}</code>.
  *
  * @author Sára Juhošová
  */
-class DoubleAssert_isNotCloseToPercentage_double_Test extends DoubleAssertBaseTest {
+class FloatAssert_isNotCloseToPercentage_float_primitive_Test extends FloatAssertBaseTest {
 
-  private final Percentage percentage = withPercentage(5.4);
-  private final double value = 62.7;
+  private final Percentage percentage = withPercentage(2.4f);
+  private final float value = 16.3f;
 
   @Override
-  protected DoubleAssert invoke_api_method() {
+  protected FloatAssert invoke_api_method() {
     return assertions.isNotCloseTo(value, percentage);
   }
 
   @Override
   protected void verify_internal_effects() {
-    verify(doubles).assertIsNotCloseToPercentage(getInfo(assertions), getActual(assertions), value, percentage);
+    verify(floats).assertIsNotCloseToPercentage(getInfo(assertions), getActual(assertions), value, percentage);
   }
 }
